+++ requirements.txt
@@ -9,7 +9,6 @@ llvmlite==0.43.0
 MarkupSafe==3.0.2
 more-itertools==10.5.0
 mpmath==1.3.0
-networkx==3.4.2
 numba==0.60.0
 numpy==2.0.2
 openai-whisper==20240930
@@ -24,4 +23,5 @@ torch==2.5.1
 tqdm==4.67.0
 typing_extensions==4.12.2
 urllib3==2.2.3
-TTS
+python-dotenv
+elevenlabs
